go router shells integration
