+++ packages/eike_app/lib/feat_navigation/eike_routes.dart
@@ -1,12 +1,8 @@
 enum EikeRoute {
-  home('/home'),
-  contact('/contact'),
-  settings('/settings'),
-  imprint('/imprint'),
-  licenses('/licenses'),
-  privacyPolicy('/privacyPolicy')
-  ;
-
-  const EikeRoute(this.route);
-  final String route;
+  home,
+  contact,
+  settings,
+  imprint,
+  licenses,
+  privacyPolicy,
 }
